+++ test/main.tf
@@ -43,4 +43,6 @@ module "mbip" {
   mbip_image_name                   = var.mbip_image_name
   mbip_release                      = var.mbip_release
   num_mbips                         = var.num_mbips
+  ssh_username                      = var.ssh_username
+  ssh_password                      = var.ssh_password
 }